[ 2.5.2 ] Minor fix & opts; ...

+++ HDC/GeneralNode.h
@@ -12,7 +12,7 @@ class CGeneralNode :
 protected:
 	vector<CData> buffer;
 	int bufferCapacity;
-	double energy;
+	int energy;
 	double energyConsumption;
